Adjust exports, gated_emit calls, time type
Fix type and API issues across WSL and provider code: remove an unnecessary i64 cast for now_unix_seconds in providers to match oauth_expires_at, change gated_emit invocations to pass app (not &app) in wsl command flows, and clean up test imports by making export_skill_dir_files unix-only. Also narrow and reorder public re-exports in infra/wsl/mod.rs (remove adapt_mcp export and reduce exported type list) to simplify the module's public API and avoid unused imports/compile errors.

+++ src-tauri/src/commands/providers.rs
@@ -945,7 +945,7 @@ pub(crate) async fn provider_oauth_fetch_limits(
         match refresh_oauth_details_for_limits(&db, &client, &details, adapter).await {
             Ok(refreshed) => details = refreshed,
             Err(err) => {
-                let now_unix = crate::shared::time::now_unix_seconds() as i64;
+                let now_unix = crate::shared::time::now_unix_seconds();
                 let still_valid = details
                     .oauth_expires_at
                     .map(|expires_at| expires_at > now_unix)
